* cygserver_ipc (class ipc_retval): Add ssize_t member to anonymous union. (ipc_retval::operator =): Add ssize_t variant for x86_64. * fhandler_floppy.cc (fhandler_dev_floppy::raw_write): Change return type to ssize_t to match fhandler.h. * fhandler_socket.cc (fhandler_socket::readv): Ditto. (fhandler_socket::write): Ditto. (fhandler_socket::writev): Ditto. (fhandler_socket::sendmsg): Ditto.

+++ winsup/cygwin/fhandler_floppy.cc
@@ -559,7 +559,7 @@ err:
   ulen = (size_t) -1;
 }
 
-int __stdcall
+ssize_t __stdcall
 fhandler_dev_floppy::raw_write (const void *ptr, size_t len)
 {
   DWORD bytes_written = 0;
